Maktab Loyhasi tugadi

+++ database/migrations/2025_05_20_124056_create_schudelis_table.php
@@ -14,7 +14,8 @@ return new class extends Migration
         Schema::create('schudelis', function (Blueprint $table) {
             $table->id();
             $table->integer('smena_id');
-            $table->integer('lesson_id');
+            $table->unsignedBigInteger('lesson_id');
+
             $table->string('week_day');
             $table->string('room');
             $table->timestamps();
